Restructured code to make code more readable and use facematch as package
